add OCR step
